Move logic to Ubiquity module
